Continuing  work on Dirichlet cells. Added class DC where we will use the spectrum of  the 13 +-1,0  lattice points' distances from the origin. Made a number of small corrections  in use of size_t, made to type int so negative numbers a allowed.
FIXED a termination error in Niggli::Reduce. Separated counters in 4 functions from using a member variable for loop counters. In Reduce with matrices, moved call to MKNorm to the end of the loop and added a new call before the loop. Basically a reinterpretation of the pseudocode  of Gruber 1973.

+++ LatticeRepresentationLib/DirichletCellFaces.h
@@ -8,7 +8,7 @@
 #include "LRL_Vector3.h"
 
 
-typedef std::vector<size_t> POINT;
+typedef std::vector<int> POINT;
 typedef std::vector<POINT> POINT_LIST; 
 typedef std::vector< std::vector<POINT>> FACES;
